requirements.txt
amqp==2.5.1
APScheduler==3.6.1
backcall==0.1.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
decorator==4.4.0
Django==2.1.1
django-common-helpers==0.9.2
django-cors-headers==3.1.1
django-crispy-forms==1.8.0
django-extensions==2.1.6
django-formtools==2.1
django-model-utils==3.2.0
django-xadmin==0.5.0
djangorestframework==3.9.2
gunicorn==19.9.0
httplib2==0.9.2
idna==2.8
importlib-metadata==0.23
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
kombu==4.6.5
more-itertools==7.2.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
pyserial==3.4
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
traitlets==4.3.3
twilio==6.32.0
tzlocal==2.0.0
urllib3==1.25.6
vine==1.3.0
wcwidth==0.1.7
Werkzeug==0.16.0
whitenoise==4.1.4
zipp==0.6.0
